+++ irohad/consensus/yac/storage/yac_vote_storage.hpp
@@ -26,12 +26,12 @@
 #include "consensus/yac/messages.hpp"  // because messages passed by value
 #include "consensus/yac/storage/storage_result.hpp"  // for Answer
 #include "consensus/yac/storage/yac_common.hpp"      // for ProposalHash
+#include "consensus/yac/storage/yac_proposal_storage.hpp"
 #include "consensus/yac/yac_types.hpp"
 
 namespace iroha {
   namespace consensus {
     namespace yac {
-      class YacProposalStorage;
 
       /**
        * Proposal outcome states for multicast propagation strategy
