+++ test/module/irohad/validation/chain_validation_test.cpp
@@ -8,7 +8,7 @@
 #include <boost/range/adaptor/indirected.hpp>
 #include "module/irohad/ametsuchi/mock_mutable_storage.hpp"
 #include "module/irohad/ametsuchi/mock_peer_query.hpp"
-#include "module/irohad/consensus/yac/yac_mocks.hpp"
+#include "module/irohad/consensus/yac/mocks/suermajority_checker.hpp"
 #include "module/shared_model/interface_mocks.hpp"
 
 using namespace iroha;
